half compute type

+++ lightseq/csrc/kernels/cuda/includes/cublas_wrappers.h
@@ -23,7 +23,7 @@ int cublas_gemm_ex(cublasHandle_t handle, cublasOperation_t transa,
 
 int cublas_gemm_ex(cublasHandle_t handle, cublasOperation_t transa,
                    cublasOperation_t transb, int m, int n, int k,
-                   const float *alpha, const float *beta, const __half *A,
+                   const __half *alpha, const __half *beta, const __half *A,
                    const __half *B, __half *C,
                    cublasGemmAlgo_t algo = CUBLAS_GEMM_DEFAULT_TENSOR_OP);
 
@@ -36,8 +36,8 @@ int cublas_strided_batched_gemm(cublasHandle_t handle, int m, int n, int k,
                                 cublasGemmAlgo_t algo = CUBLAS_GEMM_DEFAULT);
 
 int cublas_strided_batched_gemm(
-    cublasHandle_t handle, int m, int n, int k, const float *alpha,
-    const float *beta, const __half *A, const __half *B, __half *C,
+    cublasHandle_t handle, int m, int n, int k, const __half *alpha,
+    const __half *beta, const __half *A, const __half *B, __half *C,
     cublasOperation_t op_A, cublasOperation_t op_B, int stride_A, int stride_B,
     int stride_C, int batch,
     cublasGemmAlgo_t algo = CUBLAS_GEMM_DEFAULT_TENSOR_OP);
